% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{clr}
\alias{clr}
\title{clr transformation}
\usage{
clr(x)
}
\arguments{
\item{data}{a integer vector or numeric vector of non-negative values}
}
\value{
the clr-transformed vector
}
\description{
clr transformation
}
\details{

}
\examples{
clr(c(1, 2, 4))

}
