% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_trees.R
\name{color.legend}
\alias{color.legend}
\title{These functions are adapted from plotrix package and simplified here
for our purpose}
\usage{
color.legend(
  x,
  y,
  h,
  col.h,
  col.w,
  title,
  rect.col,
  legend,
  values = NULL,
  xpd = NA,
  ...
)
}
\description{
These functions are adapted from plotrix package and simplified here
for our purpose
}
