% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_trees.R
\name{color_edges}
\alias{color_edges}
\title{color_edges}
\usage{
color_edges(physeq, group = "Phylum", method = c("majority", "ace"),
  tip.only = FALSE)
}
\arguments{
\item{physeq}{Required. An instance of a \code{phyloseq} class object that has
tree slot.}

\item{group}{Optional. Either a single character string matching a variable
name in the corresponding tax_table of `physeq`, or a factor with
the same length as the number of taxa in `physeq`. Defaults to "Phylum"}

\item{method}{Optional. Ancestral group reconstruction method; either 'majority' or
'ace' (or an abbrevation of these). Defaults to "majority"}

\item{tip.only}{Optional. Logical. Should colors be computed for tips only? Defaults to FALSE.}
}
\value{
A list with components
\item{edge} A color vector for edges of \code{phy_tree{physeq}}
\item{tip} A color vector for tip labels of \code{phy_tree{physeq}}
\item{palette} The named color palette color, useful for the legend
}
\description{
Color edges and leaves of a tree according to a `group` factor. `group` should give a class
for all leaves. The class of inner nodes is inferred using majority rule or ancestral character
estimation (via ape::ace function, used with ER model)
}
\note{
The function assumes that the tree is rooted. The color palette is constructed
      automatically according to ggplot discrete hue scale
}
