% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_trees.R
\name{construct_aesthetics}
\alias{construct_aesthetics}
\title{construct_aesthetics}
\usage{
construct_aesthetics(x, color, fatten.edges.by, deviation = FALSE)
}
\arguments{
\item{x}{Required. List as returned by fattenEdges.}

\item{color}{Required. Colors used for edges and tips. If NULL (default), all
edges and tips will be black.}

\item{fatten.edges.by}{Required. Aesthetics that will be reconstructed. Subset of 'size',
'alpha' and 'color'}

\item{deviation}{Required. Logical. Used to construct an automatic color scale. If TRUE,
a divergent color scale is used, if FALSE, a sequential one.}
}
\value{
A list with components
\itemize{
\item edge.size A matrix of size nsamples(physeq) x number of edges in \code{phy_tree(physeq)}
with fattened edge widths
\item edge.color A matrix of size nsamples(physeq) x number of edges in \code{phy_tree(physeq)}
with colored edge widths
\item legend A named list with components \code{size}, \code{alpha} and \code{color}
used for constructing a legend.
}
}
\description{
Subsidiary function for plot_merged_trees. Transforms a list as returned by fattenEdges
to a list of aesthetics parameters for plot.phylo
}
