% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_methods.R
\name{correct_levels}
\alias{correct_levels}
\title{Internal function to correct levels in plot_samples}
\usage{
correct_levels(physeq, DF, map.var)
}
\description{
Internal function to correct levels in plot_samples
}
\keyword{internal}
