% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{edge_pca}
\alias{edge_pca}
\title{edge_pca}
\usage{
edge_pca(physeq, number.components = 2, method = c("normal", "sparse",
  "regularized"), number.edges = 50, ...)
}
\arguments{
\item{physeq}{Required. A \code{\link{phyloseq-class}} class object}

\item{number.components}{Optional. Numeric. Number of components used in the PCA, defaults to 2.}

\item{number.edges}{Optional. Integer. Number of non-zero loadings for each
component in the sparse PCA setting.}

\item{...}{Optional. Additional parameters passed on to sparcePca,
regularizedPca and pca, such as number of iterations and
convergence criterion.}

\item{method.}{Optional. PCA method. Any unambiguous abbreviation of 'normal',
'sparse' and 'regularized'. Defaults to 'sparse'.}
}
\value{
An \code{edge_pca} class object with components:
\itemize{
\item loadings: The loadings of the principal component axis
\item scores: The score of the samples
\item values: A date frame with components eigenvalues, relative eigenvalues and
cumulative eigenvalues.
}
}
\description{
Compute edge PCA components of a metagenomic sample
using the edge_pca idea developed in Matsen and Evans (2012)
with regularization of the edges for the PCA part as proposed in
Shen, H. and Huang, J. Z. (2008).
}
\note{
scores and extract_eigenvalues methods are associated with 'edge_pca'.
}
\examples{
data(food)
epca <- edge_pca(food)
}
\references{
Shen, H. and Huang, J. Z. (2008). Sparse principal component
analysis via regularized low rank matrix approximation. \emph{Journal
of Multivariate Analysis} \emph{99}, 1015-1034.

Matsen, F. A. and Evans, S. (2012). Edge Principal Components
and Squash Clustering: Using the Special Structure of Phylogenetic
Placement Data for Sample Comparison. \emph{Plos One} \emph{8}(3):e56859.
}
\seealso{
\code{\link{pca}}, \code{\link{sparsePca}}, \code{\link{regularizedPca}}
}
