% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_methods.R
\name{estimate_local_specificity}
\alias{estimate_local_specificity}
\title{Estimate local specificity of an OTU to a group}
\usage{
estimate_local_specificity(physeq, group, index = c("fraction",
  "indspec"), freq = TRUE, B = 999, se = TRUE, parallel = FALSE)
}
\arguments{
\item{physeq}{phyloseq class object, otu abundances are extracted from this object}

\item{group}{Either the a single character string matching a variable name in
the corresponding sample_data of `physeq`, or a factor with the same
length as the number of samples in `physeq`.}

\item{index}{Either "fraction" (default) or "indspec".
Method used for computing local specificity (see details).}

\item{freq}{Logical. Defaults to TRUE. Should counts be replaced by frequencies (TRUE)
or kept as such (FALSE). TRUE corresponding to weighting all samples equally
while FALSE weights them with their library size.}

\item{B}{Only used if se is TRUE, number of bootstrap replicates used
to compute quantiles.}

\item{se}{Logical. Defaults to TRUE. Variability of the computed specificities be assessed
(by bootstrapping samples within levels of factor). If TRUE, quantiles
5, 25, 50, 75 and 95 of specificity coefficients are returned.}

\item{parallel}{Logical, should computations be performed in parallel using \code{\link{mclapply}}
and parallel framework)}
}
\value{
A data frame with components
* specificity: observed local specificity
* group: factor level
* abundance: local relative abundance of otu (all samples of a level are weighted equally)
* mean, sd, quantiles 5, 25, 50, 75 and 95% if 'se' is TRUE
Specificity method is used as attribute 'index' of the data frame.
}
\description{
Estimate local specificity of an OTU to a group
}
\details{
Method "fraction" correspond to fraction of reads coming from a given group. Method "indspec" corresponds to fraction x local prevalence in that group.
}
\examples{
data(food)
estimate_local_specificity(food, group = "EnvType", index = "indspec", se = FALSE)
}
