% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{estimate_prevalence}
\alias{estimate_prevalence}
\title{Compute local prevalence and specificity of an OTU in different levels of a group}
\usage{
estimate_prevalence(physeq, group, format = c("long", "wide"),
  rarefy = TRUE)
}
\arguments{
\item{physeq}{phyloseq class object, otu abundances are extracted from this object}

\item{group}{Either the a single character string matching a variable name in
the corresponding sample_data of `physeq`, or a factor with the same
length as the number of samples in `physeq`.}

\item{format}{long (default) or wide. Should the result be organised as a long data.frame
with columns "otu", "group", "prevalence" and "specificity" or a numeric matrix
with otus in rows and group specificty/group prevalence in column.}

\item{rarefy}{(Logical). Default TRUE. Rarefy samples before computing prevalences.}
}
\value{
A tibble with components
* prevalence: observed prevalence in group
* specifity : specifity (prevalence / global prevalence) in group
* group: factor level
* otu: otu
or a matrix m of size `ntaxa(physeq)` times 2*N where
N is the number of levels (groups) in `group` and
* `m[i, g]` is the prevalence of otu i in group g
* `m[i, N+g]` is the specificity of otu i to group g
}
\description{
Compute local prevalence and specificity of an OTU in different levels of a group
}
\details{
Specificity is defined here as local prevalence / global prevalence,
         and therefore only deals with presence/absence data. Rarefaction is
         therefore recommended.
}
\examples{
data(food)
estimate_prevalence(food, group = "EnvType", rarefy = TRUE)
}
