% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_methods.R
\name{estimate_specificity}
\alias{estimate_specificity}
\title{Estimate specificity of an OTU to a group}
\usage{
estimate_specificity(
  physeq,
  group,
  index = c("shannon", "simpson", "yanai", "indspec"),
  freq = TRUE,
  B = 999,
  se = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{physeq}{phyloseq class object, otu abundances are extracted from this object}

\item{group}{Either the a single character string matching a variable name in
the corresponding sample_data of \code{physeq}, or a factor with the same
length as the number of samples in \code{physeq}.}

\item{index}{Specificity index used for computing specificity.}

\item{freq}{Logical. Defaults to TRUE. Should counts be replaced by frequencies (TRUE)
or kept as such (FALSE). TRUE corresponding to weighting all samples equally
while FALSE weights them with their library size.}

\item{B}{Only used if \code{se} is TRUE, number of bootstrap replicates used
to compute quantiles.}

\item{se}{Logical. Defaults to TRUE. Variability of the computed specificities be assessed
(by bootstrapping samples within levels of factor). If TRUE, quantiles
5, 25, 50, 75 and 95 of specificity coefficients are returned.}

\item{parallel}{Logical, should computations be performed in parallel using \code{\link{parallel::mclapply}}
and parallel framework)}
}
\value{
A data frame with components
\itemize{
\item specificity: observed specificity
\item group: factor level in which OTU is most abundant
\item abundance: overall relative abundance of otu (all levels are weighted equally)
\item mean, sd, quantiles 5, 25, 50, 75 and 95\% if \code{se} is TRUE
Specificity method is used as attribute \code{index} of the data frame.
}
}
\description{
Estimate specificity of an OTU to a group
}
\examples{
data(food)
estimate_specificity(food, group = "EnvType", index = "shannon", se = FALSE)
}
\seealso{
\code{\link[=estimate_local_specificity]{estimate_local_specificity()}}
}
