% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_correction.R
\name{extract_level}
\alias{extract_level}
\title{extract_level}
\usage{
extract_level(physeq)
}
\arguments{
\item{physeq}{Required. \code{\link{phyloseq}} class object with \code{tax_table} slot.}
}
\value{
A level vector, for use in \code{\link{sample_copy_number}}
}
\description{
Extract taxonomic level from a \code{\link{phyloseq}} class object for use in
\code{\link{sample_copy_number}}. The required levels are "Phylum" or "Class" within
Proteobacteria
}
\seealso{
\code{\link{sample_copy_number}}
}
