% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_methods.R
\name{fast_tax_glom}
\alias{fast_tax_glom}
\title{Fast alternative to \code{\link{tax_glom}}}
\usage{
fast_tax_glom(
  physeq,
  taxrank = rank_names(physeq)[1],
  bad_empty = c("", " ", "\\t")
)
}
\arguments{
\item{physeq}{Required. \code{\link{phyloseq-class}} object}

\item{taxrank}{A character string specifying the taxonomic level that you want to agglomerate over. Should be among the results of \code{rank_names(physeq)}. The default value is \code{rank_names(physeq)[1]}, which may agglomerate too broadly for a given experiment. You are strongly encouraged to try different values for this argument.}

\item{bad_empty}{(Optional). Character vector. Default to empty white spaces and tabs. Defines the bad/empty values that should be replaced by "Unknown".}
}
\value{
A taxonomically-agglomerated \code{\link{phyloseq-class}} object.
}
\description{
Fast alternative to \code{\link{tax_glom}}
}
\details{
fast_tax_glom differs from \code{\link{tax_glom}} in three important ways:
\itemize{
\item It is based on dplyr function and thus much faster
\item It does not preserve the phy_tree and refseq slots
\item It only preserves taxonomic ranks up to \code{taxrank} and discard all other ones.
The "archetype" (OTU name) for each group is chosen as the most abundant taxa
within that group (for compatibility with \code{\link{tax_glom}})
}
}
\examples{
data(food)
fast_tax_glom(food, "Species")
}
\seealso{
\code{\link{tax_glom}}, \code{\link{merge_taxa}}
}
