% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_trees.R
\name{fattenEdges}
\alias{fattenEdges}
\title{fattenEdges}
\usage{
fattenEdges(
  physeq,
  method = c("linear", "logarithmic"),
  width.lim = c(0.1, 4),
  base = 10,
  deviation = FALSE
)
}
\arguments{
\item{physeq}{Required. An instance of a \code{phyloseq} class object with otu_table and
phy_tree slots.}

\item{method}{Optional. Edge fattening method; either 'linear' or
'logarithmic' (or an abbrevation of these). Defaults to "linear"}

\item{width.lim}{Optional. Numeric. Minimum and maximal edge after fattening.
Defaults to c(0.1, 4). Set to c(0, x) for true linear scaling.}

\item{base}{Optional. Numeric. Base used for logarithmic scaling. Defaults to 10.}

\item{deviation}{Optional. Logical. Should the frequency (or count) of a taxa in a sample
be normalized with respect to the average of that taxa over all samples
before fattening. Defaults to FALSE. If method = "linear", the values are
centered around the mean (deviation). If method = "logarithmic", they are
divided by the mean (relative deviation)}
}
\value{
A list with components
\itemize{
\item edge.width A matrix of size nsamples(physeq) x number of edges in \code{phy_tree(physeq)}
with fattened edge widths
\item edge.raw.width A matrix of size nsamples(physeq) x number of edges in \code{phy_tree(physeq)}
with raw edge widths (untransformed counts, frequencies or deviations)
\item edge.presence A logical matrix of size nsamples(physeq) x number of edges in
\code{phy_tree(physeq)}. edge.presence\link{i, j} is TRUE if edge j is found in community i.
\item tip.presence A logical matrix of size nsamples(physeq) x ntaxa(physeq) in
tip.presence\link{i, j} is TRUE if taxa j is found in community i.
\item pendant.edges A logical vector indicating which edges are pendant (i.e lead to a tip)
\item legend A named vector featuring edge widths and labels for legend.
}
}
\description{
Functions for plotting trees with colored edges and tips according to data stored in
tax_table and otu_table of corresponding phyloseq class object
Fatten edges of a tree monotonically with the number of reads originating from that edge
in a community matrix.
}
\note{
width.lim behaves differently in "linear" and "logarithmic" modes. In linear mode, it is
used to scale edges widths. In logarithmic mode, widths are only scaled if they
go out of the range specified by \code{width.lim}.
}
