% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_phyloseq}
\alias{filter_phyloseq}
\title{Prevalence and abundance filters, applied independently on each group}
\usage{
filter_phyloseq(physeq, prev.thresh = 0.5, abund.thresh = 0.001,
  group = rep(1, nsamples(physeq)), rarefy = TRUE)
}
\arguments{
\item{physeq}{A \code{phyloseq} class object}

\item{prev.thresh}{Minimum prevalence in a group for a taxa to be kept}

\item{abund.thresh}{Minimum relative abundance in a group for a taxa to be kept}

\item{group}{Groups}

\item{rarefy}{Logical. Should samples be rarefied before abundances and prevalence are computed.}
}
\value{
A list of taxa that pass the fitlers. To be used with \code{prune_sample}
}
\description{
Prevalence and abundance filters, applied independently on each group
}
