% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{find_upper_ranks}
\alias{find_upper_ranks}
\title{extract in order all ranks above above a given rank}
\usage{
find_upper_ranks(physeq, ranks)
}
\arguments{
\item{physeq}{Required. \code{\link{phyloseq-class}} object with a taxonomic table.}

\item{ranks}{Required. One or more ranks whose ancestry are search (e.g. "Phylum", "Species", etc). Values not encountered in \code{physeq} are silently ignored}
}
\value{
a vector of rank names
}
\description{
extract in order all ranks above above a given rank
}
\examples{
data(food)
find_upper_ranks(food, "Phylum") ## c("Kingdom", "Phylum")
find_upper_ranks(food, c("Class", "Genus")) ## c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus")
}
