% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_trees.R
\name{get_legend_dimension}
\alias{get_legend_dimension}
\title{get_legend_dimension}
\usage{
get_legend_dimension(leg)
}
\arguments{
\item{leg}{Required. A list a produced by 'construct_aesthetics' in its
'legend' component}
}
\value{
Dimensions of the legend.
}
\description{
Wrapper around legend to get legend dimension. Intended for use within
plot_pretty_tree
}
\note{
This function is intended for use within plot_merged_tree, no argument checking is performed.
}
