% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{ggpdrare}
\alias{ggpdrare}
\title{Rarefaction curves for phylogenetic diversity, ggplot-style}
\usage{
ggpdrare(physeq, step = 10, label = NULL, color = NULL, log = TRUE,
  replace = FALSE, se = TRUE, plot = TRUE, parallel = FALSE)
}
\arguments{
\item{physeq}{phyloseq class object, from which abundance data are extracted}

\item{step}{Step size for sample size in rarefaction curves}

\item{label}{Default \code{NULL}. Character string. The name of the variable to map to text labels on the plot. Similar to color option but for plotting text.}

\item{color}{(Optional). Default \code{NULL}. Character string. The name of the variable to map to colors in the plot. This can be a sample variable (among the set returned by \code{\link[=sample_variables]{sample_variables(physeq)}} or taxonomic rank (among the set returned by \code{[=rank_names]{rank_names(physeq)}}). Finally, The color scheme is chosen automatically by \link{ggplot}, but it can be modified afterward with an additional layer using \link{scale_color_manual}.}

\item{log}{(Otional). Default 'TRUE'. Logical value. Should sample size
be represented using a log10 scale? Can be modified with
\code{scale_x_*} functions from \link{ggplot2}}

\item{replace}{Logical. If TRUE (default), population are treated as of infinite size,
with probabilities of occurence of a taxa computed from
the (finite size) community data.}

\item{se}{Default TRUE. Logical. Should standard errors be computed.}

\item{plot}{Logical, should the graphic be plotted.}

\item{parallel}{should rarefaction be parallelized (using parallel framework)}
}
\value{
a ggplot2 object
}
\description{
Rarefaction curves for phylogenetic diversity, ggplot-style
}
\examples{
data(food)
ggpdrare(food, se = FALSE, step = 100, color = "EnvType")
}
