% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{ggtree}
\alias{ggtree}
\title{ggtree}
\usage{
ggtree(physeq, group = NULL, freq = TRUE, method = "linear",
  base = 10, edge.group = "Phylum", edge.method = "majority",
  width.lim = c(0.1, 4), ...)
}
\arguments{
\item{physeq}{Required. An instance of a \code{phyloseq} class object that has
tree slot.}

\item{group}{Optional. Either a single character string matching a variable
name in the corresponding sam_data of \code{physeq}, or a factor with
the same length as the number of taxa in \code{physeq}. If provided, the samples
are aggregated by factor levels before being plotted}

\item{freq}{Logical. Should counts be replaced by frequencies (TRUE) or kept as such (FALSE).
Defaults to TRUE. TRUE corresponding to weighting all samples equally while FALSE
weights them with their library size.}

\item{method}{Optional. Edge fattening method; either 'linear' or
'logarithmic' (or an abbrevation of these). Defaults to "linear"}

\item{base}{Optional. Numeric. Base used for logarithmic scaling. Defaults to 10.}

\item{edge.group}{Optional. Either a single character string matching a variable
name in the corresponding tax_table of \code{physeq}, or a factor with
the same length as the number of taxa in \code{physeq}. Defaults to "Phylum"}

\item{edge.method}{Optional. Ancestral group reconstruction method; either 'majority' or
'ace' (or an abbrevation of these). Defaults to "majority"}

\item{width.lim}{Optional. Numeric. Minimum and maximal edge after fattening.
Defaults to c(0.1, 4). Set to c(0, x) for true linear scaling.}

\item{...}{Optional. Additional arguments passed on to plot.phylo to determine
the shape of the tree.}
}
\value{
A list of two data frames properly formatted for ggplot.
}
\description{
ggplot-like implementation of plot.phylo from the ape package for \code{phyloseq} class
objects. Allows more representation that ggphylo and plot_tree. Returns a data frame representing
the tree that is useful for ggplot graphical function.
}
\note{
ggtree assumes that the tree is rooted.
}
