% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{incidence_matrix}
\alias{incidence_matrix}
\title{incidence_matrix}
\usage{
incidence_matrix(phy)
}
\arguments{
\item{phy}{Required. A \code{phylo} class object}
}
\value{
An incidence matrix M of size nedges(tree) x ntaxa(tree) where
M\link{i,j} is set to 1 if taxa derives from edge i and 0 otherwise.
}
\description{
Compute incidence matrix of a tree
}
\note{
incidence_matrix assumes that the tree is rooted. If not, it roots it
it at an arbitrary taxa (taxa 1).
}
\examples{
data(food)
incidence_matrix(phyloseq::phy_tree(food))
}
