% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{incidenceMatrix}
\alias{incidenceMatrix}
\title{incidenceMatrix}
\usage{
incidenceMatrix(phy)
}
\arguments{
\item{phy}{Required. A \code{phylo} class object}
}
\value{
An incidence matrix M of size nedges(tree) x ntaxa(tree) where
        M[i,j] is set to 1 if taxa derives from edge i and 0 otherwise.
}
\description{
Compute incidence matrix of a tree
}
\note{
incidenceMatrix assumes that the tree is rooted. If not, it roots it
      it at an arbitrary taxa (taxa 1).
}
