% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse.R
\name{lefse}
\alias{lefse}
\title{Performs a LefSe analyses on the data set, based on a grouping variable given either as a sample variable or a factor.}
\usage{
lefse(physeq, group, pseudocount = 1, padj.threshold = 0.05)
}
\arguments{
\item{physeq}{Required. \code{\link{phyloseq-class}} object}

\item{group}{Required. Grouping variable, used to compute the core microbiome in a groupwise fashion. Either a single character string matching a variable name in the corresponding \code{sample_data} of \code{physeq}, or a factor with the same length as the number of samples in \code{physeq}.}

\item{pseudocount}{Optional. Pseudocount added to all counts, to avoid problems during clr transformation. Defaults to 1 and automatically replaced by 0 if all counts are between 0 and 1.}

\item{padj.threshold}{Optional. Adjusted p-value threshold used to include taxa in the LDA model.Defaults to 0.05}
}
\value{
A data frame with three columns
\itemize{
\item \code{OTU} : taxa name
\item \code{padj}: adjusted p-value (BH correction) of the Kruskall-Wallis test
\item \code{LD}  : Matrix or vector of LD scores
}
}
\description{
The purpose of this method is to reproduce the lefse code in R. Briefly:
\itemize{
\item the taxa counts are clr transformed (after adding a pseudo count of 1)
\item A univariate Kruskall-Wallis test is performed on each taxa to assess association with the grouping structure
\item A LDA model is fitted on the significant taxa
}
}
\examples{
data(food)
lefse <- lefse(food, "FoodType", padj.threshold = 0.001)
}
