% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{mclr}
\alias{mclr}
\title{mclr transformation}
\usage{
mclr(data, c = 1)
}
\arguments{
\item{data}{a integer vector or numeric vector of non-negative values}

\item{c}{positive constant used to separate the minimum value (after clr transform) and the zeros. The default value is the one recommended by the authors.}
}
\value{
the modified clr-transformed vector as defined in Yoon, Gaynanova and Müeller (2019)
}
\description{
mclr transformation
}
\details{
The mclr transform of a vector x is given by
\deqn{mclr(x) = \left(0, \frac{log(x_i)}{g(x)} + \varepsilon, \dots \right)}{%
mclr(x) = (0, log(x_i)/g(x) + epsilon, ...)}
where \eqn{g(x)} is the geometric mean of x (excluding 0s) and all non-zeros values are shifted by epsilon to preserve the original ordering of the data.
}
\examples{
counts <- matrix(c(0, 1, 1, 1, 2, 0), nrow = 2)
mclr(counts)

}
\references{
Yoon, Gaynanova and Müeller (2019), Frontiers in Genetics, Microbial Networks in SPRING - Semi-parametric Rank-Based Correlation and Partial Correlation Estimation for Quantitative Microbiome Data. doi:10.3389/fgene.2019.00516
}
