% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{merge_group}
\alias{merge_group}
\title{Merge samples based on a grouping variable given either as a sample variable or a factor.}
\usage{
merge_group(physeq, group, fun = c("sum", "mean"), update.names = TRUE)
}
\arguments{
\item{group}{(Required). Either the a single character string matching a variable name in
the corresponding sample_data of \code{x}, or a factor with the same length as
the number of samples in \code{x}.}

\item{fun}{Either "sum" (default) or "mean". Controls how abundances are merged within each factor level.}

\item{update.names}{Logical. If \code{TRUE} (default) merged samples are named according to the levels of the grouping variable. Else, keep the name of the first sample in each group.}
}
\value{
A phyloseq object that has had its sample indices merged according to the factor indicated by the group argument. The output class matches x.
}
\description{
The purpose of this method is to merge/agglomerate the sample indices of a phyloseq object according to a categorical variable contained in a sample_data or a provided factor. Unlike the original \code{\link{merge_samples}}, sample metadata are not coerced to factors. Instead, the metadata from the first sample in each group is kept and the samples are renamed according to the levels of the factors.
}
\examples{
data(food)
mg <- merge_group(food, "EnvType")
mg
## Note that sample data are preserved
phyloseq::sample_data(mg)
}
\seealso{
\code{\link{merge_samples}}
}
