% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCM_fit.R
\name{ncm_fit}
\alias{ncm_fit}
\title{Fit Sloan et al. (2006) Neutral Community Model and return results}
\usage{
ncm_fit(physeq, threshold, absolute.threshold = 1)
}
\arguments{
\item{physeq}{phyloseq class object}

\item{threshold}{Numeric. Detection threshold (in \link{0, 1}): minimum relative abundance required to detect a species in a sample. Ignored if \code{absolute_threshold} is not \code{NULL}}

\item{absolute.threshold}{Integer. Minimum count required to detect a species in a sample.}
}
\value{
A list of two tibbles and a list
\itemize{
\item a tibble called \code{ncm_prediction} containing the predicted NCM curve for the data, with 95\% confidence interval for the occupancy estimate. Contains 4 columns:
\itemize{
\item \code{freq} (frequency in the regional pool)
\item \code{occ} (predicted occupancy)
\item \code{occ_[lower|upper]} (confidence interval for the occupancy based on the Wilson confidence interval around the predicted occupancy)
}
\item a tibble called \code{observations} with one row per species containing its predicted frequency, occupancy (with confidence interval) and the observed occupancy under the fitted NCM. Contains 6 columns:
\itemize{
\item \code{otu} (otu name)
\item \code{freq} (estimated frequency in the regional pool)
\item \code{occ} (predicted occupancy)
\item \code{occ_[lower|upper]} (confidence interval for the occupancy based on the Wilson confidence interval around the predicted occupancy)
\item \code{occ_obs} (observed occupancy).
}
\item a list called \code{fit} with the estimated value for the migration parameter m and the fitted ncm curve for the data.
}
}
\description{
Fit Sloan et al. (2006) Neutral Community Model and return results
}
\examples{
data(food)
model <- ncm_fit(food)
}
