% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCM_fit.R
\name{ncm_loglik}
\alias{ncm_loglik}
\title{Compute the log-likelihood of community data under Sloan et al. (2006) Neutral Community Model}
\usage{
ncm_loglik(
  abundances,
  migration = 0.1,
  threshold = 0.01,
  absolute.threshold = NULL
)
}
\arguments{
\item{abundances}{Required. Community data (count data matrix)}

\item{migration}{Numeric. Migration rate (in \link{0, 1}) of the NCM}

\item{threshold}{Numeric. Detection threshold (in \link{0, 1}): minimum relative abundance required to detect a species in a sample. Ignored if \code{absolute_threshold} is not \code{NULL}}

\item{absolute.threshold}{Integer. Minimum count required to detect a species in a sample.}
}
\value{
A numeric value, the log-likelihood of the observed data under the NCM model.
}
\description{
Compute the log-likelihood of community data under Sloan et al. (2006) Neutral Community Model
}
\examples{
data(food)
ncm_loglik(phyloseq::otu_table(food), absolute.threshold = 1)
}
\references{
Sloan, W.T. and Lunn, M. and Woodcock, S. and Head, I.M. and Nee, S. and Curtis, T.P. (2006). Quantifying the roles of immigration and chance in shaping prokaryote community structure. Environmental Microbiology. 8:732–740. doi:10.1111/j.1462-2920.2005.00956.x.
}
