% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{pca}
\alias{pca}
\title{pca}
\usage{
pca(x, number.components = NULL, center = TRUE, scale. = FALSE)
}
\arguments{
\item{x}{Required. Numeric matrix.}

\item{number.components}{Optional. Integer. Number of components returned by the PCA.
Defaults to NULL which returns all components}

\item{center}{Optional. Logical. Should x be shifted to be (column-wise) 0-centered?
Defaults to TRUE. Alternatively, a vector of length equal to the
number of columns in x. The value is passed on to scale.}

\item{scale.}{Optional. Logical. Should the variable be normalized to have unit
variance. Defaults to TRUE. Alternatively, a vector of length
equal to the number of columns in x. The value is passed on to scale.}
}
\value{
A list with components
\itemize{
\item number.components: The number of components used
\item loadings: The matrix of variable loadings
\item scores: The matrix of sample scores
\item values: A list with components
\itemize{
\item Eigenvalues: the eigenvalues of the covariance/correlation
\item Relative_eig: the normalized eigenvalues (sum up to 1)
\item Cumul_eig: the cumulative eigenvalues (useful for scree plot)
}
\item center, scale: centering and scaling vectors, used, if any.
Otherwise, FALSE.
}
}
\description{
Perform a principal component analysis on a given data matrix using
Singular Value Decomposition.
}
\seealso{
\code{\link{edge_pca}}, \code{\link{sparsePca}}, \code{\link{regularizedPca}}
}
