% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{phylodiv}
\alias{phylodiv}
\title{Phylogenetic diversity}
\usage{
phylodiv(physeq, theta = 0)
}
\arguments{
\item{physeq}{phyloseq class object, from which phylogeny and abundance data are extracted}

\item{theta}{parameter that determines the balance in the Balance Weighted Phylogenetic Diversity. See McCoy and Matsen (2013) for details on BWPD. Faih's PD correspond to theta = 0 (default).}
}
\value{
A data.frame that collates \code{sample_data(physeq)} and an additional \code{pd} column with phylogenetic diversities.
}
\description{
Phylogenetic diversity
}
\examples{
data(food)
phylodiv(food)
}
