% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{phyloseq_to_biom}
\alias{phyloseq_to_biom}
\title{Create a \link{biom-class} from a \link{phyloseq-class} object.}
\usage{
phyloseq_to_biom(physeq, biom_format = c("frogs", "standard"))
}
\arguments{
\item{physeq}{(Required). A \link{phyloseq-class} object.}

\item{biom_format}{(Optional). Either "frogs" (default) or "standard". Controls the way the taxonomy is written in the \link{biom-class} object. \code{biom_format = "frogs"} is intended for later use with \link{import_frogs} and \code{biom_format = "standard"} for later use with \link{import_biom}.}
}
\value{
A \link{biom-class}
}
\description{
Create a \link{biom-class} from a \link{phyloseq-class} object.
}
\details{
This functions differs from \link{make_biom} in the way it encodes the taxonomy
in the 'rows' metadata. The difference is intended for better compatibility with phyloseq
\code{import_*} class of functions.
}
\examples{
\dontrun{
data(food)
phyloseq_to_biom(food)
}
}
\seealso{
\link{make_biom}, \link{write_phyloseq}
}
