% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{phyloseq_to_tsv}
\alias{phyloseq_to_tsv}
\title{Export a \link{phyloseq-class} to a tsv file}
\usage{
phyloseq_to_tsv(physeq)
}
\arguments{
\item{physeq}{(Required). A \link{phyloseq-class} object}

\item{tsv_file}{(Required). A character string indicating the file location of the tsv file.}
}
\value{
Nothing. The function is used for its side effect of exporting a \link{phyloseq-class} object to a tsv file.
}
\description{
Export a \link{phyloseq-class} to a tsv file
}
\details{
The tsv files contains:
\itemize{
\item taxa names
\item taxa affiliation (when available)
\item taxa sequences (when available)
\item taxa abundances across samples (1 column per sample)
\item taxa prevalence
\item taxa mean abundance (conditional upon presence)
All metadata pertaining to the samples is lost.
}
}
\examples{
data(food)
phyloseq_to_tsv(food)
}
