% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_methods.R
\name{plot_clust}
\alias{plot_clust}
\title{Wrapper around hclust to represent clustering tree with leaves colored according to a factor.}
\usage{
plot_clust(
  physeq,
  dist,
  method = "ward.D2",
  color = NULL,
  label = "label",
  title = paste(method, "linkage clustering tree"),
  palette = NULL,
  ...
)
}
\arguments{
\item{physeq}{phyloseq class object}

\item{dist}{distance matrix (dist class) or character to be used in phyloseq::distance function}

\item{method}{(character) linkage method used in hclust, defaults to "ward.D2"}

\item{color}{(character) variable name used to color tree leaves. Defaults to NULL}

\item{label}{(character) variable name used to label tree leaves. Defaults to "label" for sample names.}

\item{title}{(character) optional. Plot title, defaults to "method" clustering tree.}

\item{palette}{(named color vector) optional. Manual color palette}

\item{...}{(optional) additional parameters passed on to theme(axis.text.x = element_text(...)) to control label size, justification, ...}
}
\value{
Nothing, used for its plotting side effect
}
\description{
Wrapper around hclust to represent clustering tree with leaves colored according to a factor.
}
\examples{
data(food)
# Basic plot
plot_clust(food, dist = "unifrac", color = "EnvType")
# Slightly better plot
plot_clust(food, dist = "unifrac", color = "EnvType", label = "EnvType", size = 8) + ggplot2::theme(legend.position = "none")
}
