% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_methods.R
\name{plot_composition}
\alias{plot_composition}
\title{Versatile function for plotting compositions graphs}
\usage{
plot_composition(physeq, taxaRank1 = "Phylum",
  taxaSet1 = "Proteobacteria", taxaRank2 = "Family",
  numberOfTaxa = 9, fill = NULL, startFrom = 1, x = "Sample",
  y = "Abundance", sampleOrder = NULL, facet_grid = NULL, ...)
}
\arguments{
\item{physeq}{phyloseq class object}

\item{taxaRank1}{Taxonomic rank used for subsetting. If NULL, \code{taxaSet1} is ignored and no subsetting of taxa is performed.}

\item{taxaSet1}{Subset of taxa (at rank \code{taxaRank1}) to keep for the plots. If NULL, no subsetting is applied.}

\item{taxaRank2}{taxonomic level used to agglomerate, use "OTU", "ASV" or NULL to enforce no aggregation}

\item{numberOfTaxa}{number of (most abundant) taxa to keep at level taxaRank2}

\item{fill}{Taxonomic rank used for filling (should be between taxaRank1 and taxaRank2 in the hierarchy of ranks)}

\item{startFrom}{\code{startFrom - 1} is the number of (most abundant) taxa to discard before
selecting the \code{numberOfTaxa} most abundant taxa.}

\item{x}{Variable mapped to x-axis}

\item{y}{Variable mapped to y-axis}

\item{sampleOrder}{Sample order (ignored if \code{x} is different from "Samples")}

\item{facet_grid}{variable used for faceting.}

\item{...}{Additional arguments passed on to geom_bar}
}
\value{
a ggplot2 graphic
}
\description{
Versatile function for plotting compositions graphs
}
\details{
Allows the user to restrict the plot to taxa in the set \code{taxaSet1} at level \code{taxaRank1} and aggregate taxa at level \code{taxaRank2} in the plot. The plot is limited to \code{numberOfTaxa} taxa (defaults to 9) and other taxa are automatically lumped in the "Other" category.
}
\examples{
data(food)
plot_composition(food, "Kingdom", "Bacteria", "Family", fill = "Phylum", facet_grid = "EnvType")
}
