% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_methods.R
\name{plot_dist_as_heatmap}
\alias{plot_dist_as_heatmap}
\title{Plot a distance matrix as a heatmap}
\usage{
plot_dist_as_heatmap(dist, order = NULL, title = NULL,
  low = "#B1F756", high = "#132B13", show.names = FALSE)
}
\arguments{
\item{dist}{Distance matrix between the samples}

\item{order}{Optional. Sample order used for plotting}

\item{title}{Optional. Plot title.}

\item{low}{Optional. Default "#B1F756". Color for low values}

\item{high}{Optional. Default "#132B13". Color for high values}

\item{show.names}{Optional. Default FALSE. Show sample names in heatmap. Use only
if there are not too many samples.}
}
\value{
A ggplot2 object
}
\description{
Plot a distance matrix as a heatmap
}
\examples{
data(food)
dist.bc <- distance(food, "bray")
plot_dist_as_heatmap(dist.bc)
}
