% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{plot_edge_pca}
\alias{plot_edge_pca}
\title{plot_edge_pca}
\usage{
plot_edge_pca(physeq, epca, axes = c(1, 2), widths = c(1, 1),
  heights = c(1, 1), args.loadings = NULL, ...)
}
\arguments{
\item{physeq}{Required. A \code{\link{phyloseq-class}} object}

\item{epca}{Required. A \code{\link{edge_pca}} class object}

\item{axes}{Optional. Ordination axes. Defaults to c(1, 2).}

\item{widths}{Optional. A vector of 2 values for the widths of columns. Defaults to c(1, 1)}

\item{heights}{Optional. A vector of 2 values for the heights of rows. Defaults to c(1, 1)}

\item{args.loadings}{Optional. Additional passed on to plotLoadings. Defaults to NULL}

\item{...}{Optional. Additional argument passed on to plot_ordination}

\item{type}{Required. Unambiguous abbreviation of 'variables' or 'samples'.
For compatibility with vegan and phyloseq, 'variables' can be replaced
by 'species' or 'taxa' and 'samples' can be replaced by 'sites'.}
}
\value{
Nothing. The function is called for its side effect.
}
\description{
Method for plotting an edge_pca object with enhanced axes. Expands
\code{\link{plot_loadings}} and \code{\link{plot_ordination}}
}
\examples{
data(food)
epca <- edge_pca(food)
plot_edge_pca(food, epca)
}
\seealso{
\code{\link{pca}}, \code{\link{edge_pca}}, \code{\link{regularizedPca}}, \code{\link{sparsePca}}
}
