% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{plot_loadings}
\alias{plot_loadings}
\title{plot_loadings}
\usage{
plot_loadings(
  physeq,
  epca,
  choices,
  fatten.by = c("size"),
  fatten.tips = FALSE,
  width.lim = c(0.1, 4),
  color = c("#EF8A62", "#67A9CF"),
  color.tip.by = NULL,
  missing.color = NULL,
  ...
)
}
\arguments{
\item{physeq}{Required. A \code{\link{phyloseq-class}} class instance.}

\item{epca}{Required. An \code{\link{edge_pca}} class object}

\item{choices}{Required. Ordination axes.}

\item{fatten.tips}{Optional. Logical. Should tips be fattened like pendant edges.
Defaults to FALSE.}

\item{width.lim}{Optional. Numeric. Minimum and maximal edge after fattening.
Defaults to c(0.1, 4). Set to c(0, x) for true linear scaling.}

\item{color}{Optional. Color vector of length 2 used to distinguish positive
loading edges (color\link{1}) and negative loading edges (color\link{2}).
Defaults to c("#EF8A62", "#67A9CF")}

\item{color.tip.by}{Optional. Either a single character string matching a variable
name in the corresponding tax_table of \code{physeq}, or a factor with
the same length as the number of taxa in \code{physeq}. If NULL, nothing happens.}

\item{missing.color}{Optional. Color used for tips and edges with 0 loading.
Defaults to NULL. If NULL, nothing happens. Use "white", "transparent",
or par("bg") to remove them from plot.}

\item{...}{Additional arguments passed on to plot.phylo}

\item{fatten.edges.by}{Required. Aesthetics that will be used for fattening.
Subset of 'size' and 'alpha'.}
}
\value{
Nothing, this function is used for its side effect of plotting a tree
}
\description{
Function for plotting a tree with edge fattened and colored according to
the loadings of an edge_pca results.
}
\seealso{
\code{\link{edge_pca}}
}
