% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_trees.R
\name{plot_merged_trees}
\alias{plot_merged_trees}
\title{plot_merged_trees}
\usage{
plot_merged_trees(physeq, group, freq = TRUE, method = "linear",
  missing.color = "gray", show.missing.tip = TRUE,
  fatten.edges.by = c("size"), fatten.tips = FALSE,
  color.edge.by = NULL, color.edge.method = "majority", base = 10,
  width.lim = c(0.1, 4), deviation = FALSE,
  legend.title = deparse(substitute(group)), ...)
}
\arguments{
\item{physeq}{Required. An instance of a \code{phyloseq} class object that has
tree slot.}

\item{group}{Required. Either a single character string matching a variable
name in the corresponding sam_data of `physeq`, or a factor with
the same length as the number of taxa in `physeq`.}

\item{freq}{Logical. Should counts be replaced by frequencies (TRUE) or kept as such (FALSE).
Defaults to TRUE. TRUE corresponding to weighting all samples equally while FALSE
weights them with their library size.}

\item{method}{Optional. Edge fattening method; either 'linear' or
'logarithmic' (or an abbrevation of these). Defaults to "linear"}

\item{missing.color}{Optional. Color used for tips and edges not present in the sample.
Defaults to "gray". If NULL, nothing happens. Use "white", "transparent",
or par("bg") to remove them from plot.}

\item{show.missing.tip}{Optional. Logical. Should tips not present in the sample be still be
labelled. If FALSE, corresponding labels are replaced with ""}

\item{fatten.edges.by}{Optional. Aesthetics used for edge "fattening", vector with elements in
'size', 'alpha', 'color'. Defaults to 'size' only.}

\item{fatten.tips}{Optional. Logical. Should tips labels be "fattened" as edges.
Defaults to FALSE.}

\item{color.edge.by}{Optional. A single character string matching a variable. If NULL,
nothing happens.
name in the corresponding tax_table of `physeq`, or a factor with
the same length as the number of taxa in `physeq`.
Defaults to NULL, corresponding to no color.}

\item{color.edge.method}{Optional. Ancestral group reconstruction method; either 'majority' or
'ace' (or an abbrevation of these). Defaults to "majority"}

\item{base}{Optional. Numeric. Base used for logarithmic scaling. Defaults to 10.}

\item{width.lim}{Optional. Numeric. Minimum and maximal edge after fattening.
Defaults to c(0.1, 4). Set to c(0, x) for true linear scaling.}

\item{deviation}{Optional. Logical. Should the frequency (or count) of a taxa in a sample
be normalized with respect to the average of that taxa over all samples
before fattening. Defaults to FALSE. If method = "linear", the values are
centered around the mean (deviation). If method = "logarithmic", they are
divided by the mean (relative deviation)}

\item{legend.title}{Optional. Character string used for legend title. Defaults to
`group`.}

\item{...}{Optional. Additional arguments passed on to plot.phylo.}
}
\value{
The function is mainly called for its side effect of plotting a tree
        with colored edges and leaves.
}
\description{
Merge samples according to group factor and plot taxa diversity for each categrory of the group.
}
\note{
The function assumes that the tree is rooted.
}
