% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCM_fit.R
\name{plot_ncm}
\alias{plot_ncm}
\title{Plot a ncm_fit result}
\usage{
plot_ncm(ncm.fit)
}
\arguments{
\item{ncm.fit}{scnm fit, as produced by \code{\link[=ncm_fit]{ncm_fit()}}}
}
\value{
a ggplot object
}
\description{
Plot a ncm_fit result
}
\examples{
data(food)
model <- ncm_fit(food)
plot_ncm(model)
}
