% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_trees.R
\name{plot_pretty_legend}
\alias{plot_pretty_legend}
\title{plot_pretty_legend}
\usage{
plot_pretty_legend(x, y, leg)
}
\arguments{
\item{x}{Required. x-coordinate around which the legend is
centered.}

\item{y}{Required. y-coordinate around which the legend is
centered.}

\item{leg}{Required. A list as produced by 'construct_aesthetics' in its
'legend' component}
}
\value{
Nothing. This function is used for its side effect of plotting
a legend.
}
\description{
Wrapper around legend that constructs automatic legend for the graph.
Intended for use within plot_merged_trees
}
\note{
This function is intended for use within plot_merged_tree, no argument checking is performed.
}
