% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_trees.R
\name{plot_pretty_tree}
\alias{plot_pretty_tree}
\title{plot_pretty_tree}
\usage{
plot_pretty_tree(tree, edge.size, edge.color, tip.size, tip.color,
  fatten.edges.by, fatten.tips, color.tip, ...)
}
\arguments{
\item{tree}{Required. A \code{phylo} class object}

\item{edge.size}{Required. An edge expansion vector, potentially passed on to plot.phylo}

\item{edge.color}{Required. A color vector potentially passed on to plot.phylo}

\item{tip.color}{Required. A color vector potentially passed on to plot.phylo}

\item{fatten.edges.by}{Required. Aesthetics used for edge "fattening", vector with elements in
'size', 'alpha', 'color'. Defaults to 'size' only. Determines which
of 'edge.color', 'edge.size' are passed on to
plot.phylo}

\item{fatten.tips}{Required. Aesthetics used for tip "fattening", vector with elements in
'size', 'alpha', 'color'. Defaults to 'size' only. Determines which
of 'tip.color', 'tip.size' are passed on to plot.phylo}

\item{color.tip}{Optional. Logical. Should tips be colored even if fatten.tips is set to false.}

\item{...}{Optional. Additional arguments passed on to plot.phylo.}

\item{tip.size.}{Required. A tip expansion vector, potentially passed on to plot.phylo}
}
\value{
Nothing. This function is used for its side effect of plotting a tree.
}
\description{
Wrapper around plot.phylo to plot tree with precomputed aesthetics.
Only aesthetics specified in fatten.edges.by fed to plot.phylo. The goal
of this function is to allow the user to feed plot.phylo with custom
aesthetics in addition to those automatically computed by plot_merged_trees
and use only the custom ones or the automatically computed ones.
}
\note{
This function is intended for use within plot_merged_tree, no argument checking is performed.
}
