% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rarecurve2}
\alias{rarecurve2}
\title{Build rarefaction curves from a count matrice}
\usage{
rarecurve2(
  physeq,
  step = 1,
  sample,
  xlab = "Sample Size",
  ylab = "Number of species",
  label = TRUE,
  col = "black",
  ...
)
}
\arguments{
\item{physeq}{A phyloseq class object}

\item{step}{Step size for sample sizes in rarefaction curves. Defaults to 1.}

\item{sample}{Subsample size for rarefying community, either a single value or a vector.}

\item{xlab}{Axis labels in plots of rarefaction curves.}

\item{ylab}{Axis labels in plots of rarefaction curves.}

\item{label}{Label rarefaction curves by rownames of x (logical).}

\item{col}{Custom colours for line rarefactions curves.}

\item{...}{parameters passed to ordilabels}
}
\description{
Warning, this function is deprecated and will be phased out at some point.
}
\examples{
\dontrun{
data(food)
rarecurve2(food, step = 100)
}
}
