% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{regularizedPca}
\alias{regularizedPca}
\title{regularizedPca}
\usage{
regularizedPca(x, number.components = 10, center = TRUE,
  scale. = FALSE)
}
\arguments{
\item{x}{Required. Numeric matrix.}

\item{number.components}{Optional. Integer. Number of components used for denoising.
Defaults to 10 and is automatically reduced to max(dim(x)) - 1
if 10 is too high.}

\item{center}{Optional. Logical. Should x be shifted to be (column-wise) 0-centered?
Defaults to TRUE. Alternatively, a vector of length equal to the
number of columns in x. The value is passed on to scale.}

\item{scale.}{Optional. Logical. Should the variable be normalized to have unit
variance. Defaults to TRUE. Alternatively, a vector of length
equal to the number of columns in x. The value is passed on to scale.}
}
\value{
A list with components
\itemize{
\item number.components: The number of components used
\item loadings: The matrix of variable loadings
\item scores: The matrix of sample scores
\item values: A list with components
\itemize{
\item Eigenvalues: the eigenvalues of the covariance/correlation
\item Relative_eig: the normalized eigenvalues (sum up to 1)
\item Cumul_eig: the cumulative eigenvalues (useful for scree plot)
}
\item center, scale: centering and scaling vectors, used, if any.
Otherwise, FALSE.
}
}
\description{
Perform a regularized principal component analysis on a given data matrix using
Singular Value Decomposition and regularization.
}
\seealso{
\code{\link{pca}}, \code{\link{edge_pca}}, \code{\link{regularizedPca}}
}
