% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_correction.R
\name{sample_copy_number}
\alias{sample_copy_number}
\title{sample_copy_number}
\usage{
sample_copy_number(level, dispersion = FALSE)
}
\arguments{
\item{level}{Required. Character. Taxonomic level for which the CN is required.}

\item{dispersion.}{Optional. Logical. Should the returned CN be the average within
the level or sampled according to the distribution within that level.
Distribution is approximated by a normal with mean and sd set to those
observed in the sample database. Defaults to FALSE.}
}
\value{
A numeric vector corresponding to imputed CN.
}
\description{
Use information from Vetroský and Baldrian (2013) to correct for 16S rRNA copy number variation.
Given a vector of taxonomic level (Phylum or Class within Proteobacteria), returns the average copy
number (CN) for each level. The average are taken from the database reconstructed by the authors.
}
\note{
The copy number is truncated to the [1, 15] range observed in the database. Levels
      not observed in the database have CN set to 4.2 or sample from \code{rnorm(1, 4.2, 2.7)}
      which correspond to the global mean and sd in the database.
}
\references{
Vetroský, T. and Baldrian, P.(2013). The variability of the 16S rRNA Gene in Bacterial
            Genomes and Its Consequences for Bacterial Community Analyses. _Plos One_ *8*(2):e57923
}
\seealso{
\code{\link{sample_copy_number_internal}}, \code{\link{extract_level}}
}
