% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_correction.R
\name{scale_sample_counts}
\alias{scale_sample_counts}
\title{scale_sample_counts}
\usage{
scale_sample_counts(physeq, scaling = "Phylum", dispersion = FALSE)
}
\arguments{
\item{physeq}{Required. \code{\link{phyloseq}} class object}

\item{scaling}{Optional. Either a numeric scaling factor or any
of "Phylum" or "Abundance". If "Phylum", The scaling is reconstructed
as the average copy number within "Phylum" or "Class" for Proteobacteria
using \code{\link{sample_copy_number}}. If "Abundance", counts are moderated
within each sample to shrink them towards evenness. See Details.}

\item{dispersion}{Optional. Logical. Should the returned CN be the average within
the level or sampled according to the distribution within that level.
Distribution is approximated by a normal with mean and sd set to those
observed in the sample database. Defaults to FALSE.}
}
\value{
a \code{\link{phyloseq}} class objet with scaled \code{otu_table} slot.
}
\description{
Scale sample counts from a \code{\link{phyloseq}} class object according to a
scaling vector or using the Phylum (or Class within Proteobacteria) average copy number.
Average copy number are taken from Vetroský and Baldrian (2013).
}
\note{
The copy number inferred under "Phylum" scaling is truncated to the [1, 15] range observed
      in the database. Levels not observed in the database have CN set to 4.2 or sample from
      \code{rnorm(1, 4.2, 2.7)} which correspond to the global mean and sd in the database.
      Under "Abundance" scaling, counts are moderated within each sample by assuming that the most
      abundant taxa has CN of 15, the least abundant has CN of 1 and linear interpolation between
      these extremes. This effectively evens bacterial counts (as opposed to 16S rRNA gene counts).
}
\references{
Vetroský, T. and Baldrian, P.(2013). The variability of the 16S rRNA Gene in Bacterial
            Genomes and Its Consequences for Bacterial Community Analyses. _Plos One_ *8*(2):e57923
}
\seealso{
\code{\link{sample_copy_number}}, \code{\link{extract_level}}
}
