% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{scores.pca}
\alias{scores.pca}
\title{scores.pca}
\usage{
\method{scores}{pca}(pca, choices, display = c("sites", "species",
  "samples", "variables", "taxa"), ...)
}
\arguments{
\item{pca}{Required. A \code{\link{pca}} class object}

\item{choices}{Required. Ordination axes. If missing, returns all axes.}

\item{display}{Required. Unambiguous abbreviation of 'variables' or 'samples'.
For compatibility with vegan and phyloseq, 'variables' can be replaced
by 'species' or 'taxa' and 'samples' can be replaced by 'sites'.}

\item{...}{Optional. Additional argument passed on from scores.default. Unused}
}
\value{
A matrix of scores for axes in 'choices'
}
\description{
S3 Method for extracting scores from pca object
}
\seealso{
\code{\link{pca}}, \code{\link{edge_pca}}, \code{\link{regularizedPca}}, \code{\link{sparsePca}},
}
