% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{sparsePca}
\alias{sparsePca}
\title{sparsePca}
\usage{
sparsePca(x, number.components = 10, number.variables = rep(ncol(x),
  number.components), center = TRUE, scale. = FALSE, iter.max = 500,
  tol = 1e-09)
}
\arguments{
\item{x}{Required. Numeric matrix.}

\item{number.components}{Optional. Integer. Number of components used for denoising.
Defaults to 10 and is automatically reduced to max(dim(x)) - 1
if 10 is too high.}

\item{number.variables}{Optional. Integer vector of length number.components specifying
the number of non-zero loadings for each component. By default, all
loadings can be different from 0 (no sparsity).}

\item{center}{Optional. Logical. Should x be shifted to be (column-wise) 0-centered?
Defaults to TRUE. Alternatively, a vector of length equal to the
number of columns in x. The value is passed on to scale.}

\item{scale.}{Optional. Logical. Should the variable be normalized to have unit
variance. Defaults to TRUE. Alternatively, a vector of length
equal to the number of columns in x. The value is passed on to scale.}

\item{iter.max}{Optional. Integer. The maximum number of iterations used in NIPALS for each
each component. Defaults to 500.}

\item{tol}{Optional. A (small) positive numeric used to check convergence of loadings
for each component. Defaults to 1e-09.}
}
\value{
A list with components
- \item{number.components}: The number of components used
- \item{loadings}: The matrix of variable loadings
- \item{scores}: The matrix of sample scores
- \item{values}: A list with components
    - \item{Eigenvalues}: the eigenvalues of the covariance/correlation
    - \item{Relative_eig}: the normalized eigenvalues (sum up to 1)
    - \item{Cumul_eig}: the cumulative eigenvalues (useful for scree plot)
- \item{center, scale}: centering and scaling vectors, used, if any.
                        Otherwise, FALSE.
}
\description{
Perform a sparse principal component analysis on a given data matrix using
NIPALS algorithm.
}
\seealso{
\code{\link{pca}}, \code{\link{edgePCA}}, \code{\link{regularizedPca}},
}
