% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{staggered_tax_glom}
\alias{staggered_tax_glom}
\title{Staggered taxonomic agglomeration}
\usage{
staggered_tax_glom(physeq, atomic_taxa, taxrank)
}
\arguments{
\item{physeq}{Required. \code{\link{phyloseq-class}} object}

\item{atomic_taxa}{A character string specificying the taxa that won't be agglomerated. Note that taxa can be mixed (i.e. Species and Phylum) but are not allowed to overlap.}

\item{taxrank}{A character string specifying the taxonomic level that you want to agglomerate over (except for the taxa in \code{atomic_taxa}).}
}
\value{
A taxonomically-agglomerated \code{\link{phyloseq-class}} object.
}
\description{
Staggered taxonomic agglomeration
}
\details{
staggered_tax_glom differs from \code{\link[=fast_tax_glom]{fast_tax_glom()}} by preserving some taxa during the taxonomic agglomeration phase
}
\examples{
data(food)
staggered_tax_glom(food, atomic_taxa = "BS11 gut group", taxrank = "Phylum") |> tax_table()
staggered_tax_glom(food, atomic_taxa = c("BS11 gut group", "Serratia"), taxrank = "Phylum")  |> tax_table()
}
