% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgePCA.R
\name{support_taxa}
\alias{support_taxa}
\title{support_taxa}
\usage{
support_taxa(
  physeq,
  epca,
  choices = seq(epca$number.components),
  threshold = 0,
  number = ntaxa(physeq)
)
}
\arguments{
\item{physeq}{Required. A \code{\link{phyloseq-class}} class instance.}

\item{epca}{Required. An \code{\link{edge_pca}} class object}

\item{choices}{Optional. Ordination axis for which support taxa are extracted.
Defaults to all axis.}

\item{threshold}{Optional. Numeric. Threshold that a taxa loading
must pass to be extracted. Defaults to 0.}

\item{number}{Optional. Integer. Number of taxa to extract from each axis.
Defaults to all taxa.}
}
\value{
A character vector ot taxa names.
}
\description{
Extract supporting taxa from an edge pca
}
\note{
If number and threshold are specified, extracts at most \code{number} taxa whose loadings
exceed threshold for each axis.
}
