% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_spread.R
\name{tax_spread}
\alias{tax_spread}
\title{tax_spread
Spread taxonomy according to last knowing taxa}
\usage{
tax_spread(physeq, pattern = "^NA$|Multi-affiliation|Unknown")
}
\arguments{
\item{physeq}{(Required). phyloseq-class.}

\item{pattern}{string. Pattern matching with taxon to rename. Can use regex, use \code{|} for separator}
}
\value{
physeq object with correct name
}
\description{
tax_spread
Spread taxonomy according to last knowing taxa
}
\examples{
data(food)
tax_spread(food)
}
