% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_methods.R
\name{test_local_specificity}
\alias{test_local_specificity}
\title{Assess significance of observed local specificities using permutation tests}
\usage{
test_local_specificity(
  physeq,
  group,
  index = c("fraction", "indspec"),
  freq = TRUE,
  replace = FALSE,
  type = c("local", "global"),
  B = 999,
  parallel = FALSE
)
}
\arguments{
\item{physeq}{phyloseq class object, otu abundances are extracted from this object}

\item{group}{Either the a single character string matching a variable name in
the corresponding sample_data of \code{physeq}, or a factor with the same
length as the number of samples in \code{physeq}.}

\item{index}{Specificity index used for computing specificity.}

\item{freq}{Logical. Defaults to TRUE. Should counts be replaced by frequencies (TRUE)
or kept as such (FALSE). TRUE corresponding to weighting all samples equally
while FALSE weights them with their library size.}

\item{replace}{Logical. Should samples be replaced when randomizing samples within group levels.}

\item{type}{Unambiguous abbreviation of "global" or "local". If "local", specificity of a taxa to an environment
is compared to random specificities in that environment only. If "global", it is compared to
the maximum specificity over all environments. "global" is more conservative than "local". Defaults to "local".}

\item{B}{Only used if \code{se} is TRUE, number of bootstrap replicates used
to compute quantiles.}

\item{parallel}{Logical, should computations be performed in parallel using \code{\link{parallel::mclapply}}
and parallel framework)}
}
\value{
A data frame with components
\itemize{
\item specificity: observed local specificity
\item level: factor level in which otu is most abundant
\item abundance: overall relative abundance of otu (all levels are weighted equally)
\item rawp:  raw p-value
\item adjp: adjusted p-value (computed using \code{\link[=p.adjust]{p.adjust()}} with \code{method = "fdr"})
Specificity method is used as attribute \code{index} of the data frame.
}
}
\description{
Assess significance of observed local specificities using permutation tests
}
\examples{
data(food)
test_local_specificity(food, group = "EnvType", index = "indspec")
}
\seealso{
\code{\link[=estimate_specificity]{estimate_specificity()}}, \code{\link[=estimate_local_specificity]{estimate_local_specificity()}}, \code{\link[=test_specificity]{test_specificity()}}
}
