% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_methods.R
\name{top_taxa}
\alias{top_taxa}
\title{Find the most abundant taxa at a given taxonomic rank}
\usage{
top_taxa(physeq, taxaRank = NULL, numberOfTaxa = 9)
}
\arguments{
\item{physeq}{phyloseq class object}

\item{taxaRank}{taxonomic level used for agglomeration. Default NULL, equivalent to \code{taxaRank = "OTU"} and no agglomeration.}

\item{numberOfTaxa}{Number of top taxa to return}
}
\value{
A matrix corresponding to the \code{numberOfTaxa} most abundant taxa at level \code{taxaRank}-level
}
\description{
Find the most abundant taxa at a given taxonomic rank
}
\examples{
data(food)
top_taxa(food, "Phylum", 10)
}
