% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_trees.R
\name{tree_colored_by}
\alias{tree_colored_by}
\title{tree_colored_by}
\usage{
tree_colored_by(physeq, group = "Phylum", method = c("majority",
  "ace"), legend.title = deparse(substitute(group)),
  legend.pos = "bottomright", plot = TRUE, ...)
}
\arguments{
\item{physeq}{Required. An instance of a \code{phyloseq} class object that has
tree slot.}

\item{group}{Optional. Either a single character string matching a variable
name in the corresponding tax_table of `physeq`, or a factor with
the same length as the number of taxa in `physeq`. Defaults to "Phylum"}

\item{method}{Optional. Ancestral group reconstruction method; either 'majority' or
'ace' (or an abbrevation of these). Defaults to "majority"}

\item{legend.title}{Optional. Character string used for legend title. Defaults to
`group`.}

\item{legend.pos}{Optional. keyword used for legend positon. Defaults to "bottomright"
and can take any value accepted by legend. "none" removes legend.}

\item{plot}{Optional. Logical. Should the tree be plotted or not.}

\item{...}{Optional. Additional arguments passed on to plot.phylo.}
}
\value{
The function is mainly called for its side effect of plotting a tree
        with colored edges and leaves.
}
\description{
Plot tree with edges colored by `group`. `group` should give a class for all leaves.
The class of inner nodes is inferred using majority rule or ancestral character
estimation (via ape::ace function, used with ER model)
}
\note{
The function assumes that the tree is rooted.
}
