% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{write_biom_safe}
\alias{write_biom_safe}
\title{Safe replacement of \code{\link[biomformat:write_biom]{biomformat::write_biom()}}}
\usage{
write_biom_safe(x, biom_file)
}
\description{
Works for biom data with only one column (sample), unlike the original version.
}
\examples{
data(food)
small_food <- food \%>\% prune_samples(sample_names(.)[1], . ) \%>\% prune_taxa(taxa_names(.)[1:5], .)
small_biom <- small_food \%>\% phyloseq_to_biom()
tmp_biom <- tempfile()
write_biom_safe(small_biom, tmp_biom)
readLines(tmp_biom)
}
