% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{write_phyloseq}
\alias{write_phyloseq}
\title{Export a \link{phyloseq-class} to several text files: biom, newick and fasta.}
\usage{
write_phyloseq(
  physeq,
  biom_file,
  tree_file = NULL,
  fasta_file = NULL,
  biom_format = c("frogs", "standard")
)
}
\arguments{
\item{physeq}{(Required). A \link{phyloseq-class} object}

\item{biom_file}{(Required). A character string indicating the file location of the biom formatted file.}

\item{tree_file}{(Optional). A character string indicating the file location of the newick formatted file. If \code{NULL} (default), the \code{\link[=phy_tree]{phy_tree()}} component of \code{physeq} is ignored.}

\item{fasta_file}{(Optional). A character string indicating the file location of the fasta formatted file. If \code{NULL} (default), the \code{\link[=refseq]{refseq()}} component of \code{physeq} is ignored.}

\item{biom_format}{(Optional). Either "frogs" (default) or "standard". Controls the way the taxonomy is written in the \link{biom-class} object. \code{biom_format = "frogs"} is intended for later use with \link{import_frogs} and \code{biom_format = "standard"} for later use with \link{import_biom}.}
}
\value{
Nothing. The function is used for its side effect of exporting a \link{phyloseq-class} object to text files.
}
\description{
Export a \link{phyloseq-class} to several text files: biom, newick and fasta.
}
\examples{
data(food)
tmp_biom <- tempfile()
tmp_tree <- tempfile()
write_phyloseq(food, biom_file = tmp_biom, tree_file = tmp_tree)
## The output biom can be read again as a phyloseq object
import_frogs(tmp_biom, tmp_tree)
}
