% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_methods.R
\name{ggformat}
\alias{ggformat}
\title{Format phyloseq data for easy composition plots}
\usage{
ggformat(
  physeq,
  taxaRank1 = "Phylum",
  taxaSet1 = "Proteobacteria",
  taxaRank2 = "Family",
  fill = NULL,
  numberOfTaxa = 9,
  startFrom = 1
)
}
\arguments{
\item{physeq}{phyloseq class object}

\item{taxaRank1}{Taxonomic rank used for subsetting. If NULL, \code{taxaSet1} is ignored and no subsetting of taxa is performed.}

\item{taxaSet1}{Subset of taxa (at rank \code{taxaRank1}) to keep for the plots. If NULL, no subsetting is applied.}

\item{taxaRank2}{taxonomic level used to agglomerate, use "OTU", "ASV" or NULL to enforce no aggregation}

\item{fill}{Taxonomic rank used for filling (should be between taxaRank1 and taxaRank2 in the hierarchy of ranks)}

\item{numberOfTaxa}{number of (most abundant) taxa to keep at level taxaRank2}

\item{startFrom}{\code{startFrom - 1} is the number of (most abundant) taxa to discard before
selecting the \code{numberOfTaxa} most abundant taxa.}
}
\value{
A ggplot friendly data frame with the relative abundance of the top \code{NumberOfTaxa} OTUs (starting from the \code{startFrom}th most abundant) at the taxaRank2 level within taxaSet1 at taxaRank1 level
}
\description{
Format phyloseq data for easy composition plots
}
\examples{
data(food)
ggformat(food)
}
