% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{merge_group}
\alias{merge_group}
\title{Merge samples based on a grouping variable given either as a sample variable or a factor.}
\usage{
merge_group(physeq, group, fun = mean)
}
\arguments{
\item{group}{(Required). Either the a single character string matching a variable name in
the corresponding sample_data of \code{x}, or a factor with the same length as
the number of samples in \code{x}.}

\item{fun}{(Optional). The function that will be used to merge the values that
correspond to the same group for each variable. It must take a numeric vector
as first argument and return a single value. Default is \code{\link[base]{mean}}.
Note that this is (currently) ignored for the otu_table, where the equivalent
function is \code{\link[base]{sum}}, but evaluated via \code{\link[base]{rowsum}}
for efficiency.}
}
\value{
A phyloseq object that has had its sample indices merged according to the factor indicated by the group argument. The output class matches x.
}
\description{
The purpose of this method is to merge/agglomerate the sample indices of a phyloseq object according to a categorical variable contained in a sample_data or a provided factor. Unlike the original \code{\link{merge_samples}}, sample metada are not coerced to factors. Instead, the metadata from the first sample in each group is kept.
}
\examples{
data(food)
merge_group(food, "EnvType")
## Note that sample data are preserved
sample_data(food)
}
\seealso{
\code{\link{merge_samples}}
}
