% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_methods.R
\name{plot_clust}
\alias{plot_clust}
\title{Wrapper around hclust to represent clustering tree with leaves colored according to a factor.}
\usage{
plot_clust(physeq, dist, method = "ward.D2", color = NULL,
  label = NULL, title = paste(method, "linkage clustering tree"),
  palette = NULL)
}
\arguments{
\item{physeq}{phyloseq class object}

\item{dist}{distance matrix (dist class) or character to be used in phyloseq::distance function}

\item{method}{(character) linkage method used in hclust, defaults to "ward.D2"}

\item{color}{(character) variable name used to color tree leaves. Defaults to NULL}

\item{label}{(character) variable name used to label tree leaves. Defaults to NULL}

\item{title}{(character) optional. Plot title, defaults to "method" clustering tree.}

\item{palette}{(named color vector) optional. Manual color palette}
}
\value{
Nothing, used for its plotting side effect
}
\description{
Wrapper around hclust to represent clustering tree with leaves colored according to a factor.
}
\examples{
data(food)
plot_clust(food, dist = "unifrac", color = "EnvType")
}
