% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_methods.R
\name{plot_composition}
\alias{plot_composition}
\title{Title}
\usage{
plot_composition(physeq, taxaRank1 = "Phylum",
  taxaSet1 = "Proteobacteria", taxaRank2 = "Family",
  numberOfTaxa = 9, fill = NULL, startFrom = 1, x = "Sample",
  y = "Abundance", facet_grid = NULL)
}
\arguments{
\item{physeq}{phyloseq class object}

\item{taxaRank1}{taxonomic level in which to do the first subsetting}

\item{taxaSet1}{subset of level taxaRank1 to use}

\item{taxaRank2}{taxonomic level used to agglomerate}

\item{numberOfTaxa}{number of (most abundant) taxa to keep at level taxaRank2}

\item{fill}{Taxonomic rank used for filling (should be between taxaRank1 and taxaRank2 in the hierarchy of ranks)}

\item{startFrom}{\code{startFrom - 1} is the number of (most abundant) taxa to discard before
selecting the `numberOfTaxa` most abundant taxa.}

\item{x}{Variable mapped to x-axis}

\item{y}{Variable mapped to y-axis}

\item{facet_grid}{variable used for faceting.}
}
\value{
a ggplot2 graphic
}
\description{
Title
}
\details{
Allows the user to restrict the plot to taxa in the set `taxaSet1` at level `taxaRank1` and aggregate taxa at level `taxaRank2` in the plot. The plot is limited to `numberOfTaxa` taxa (defaults to 9) and other taxa are automatically lumped in the "Other" category.
}
\examples{
data(food)
plot_composition(food, "Kingdom", "Bacteria", "Family", fill = "Phylum", facet_grid = "EnvType")
}
