% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_spread.R
\name{tax_spread}
\alias{tax_spread}
\title{tax_spread}
\usage{
tax_spread(physeq, pattern = "^NA$|Multi-affiliation|Unknown", explicit = TRUE)
}
\arguments{
\item{physeq}{(Required). phyloseq-class.}

\item{pattern}{string. Pattern matching with taxon to rename. Can use regex, use \code{|} for separator}

\item{explicit}{Logical. Should the spreading be explicit (e.g. "Unknown Firmicute species" for the species rank) (TRUE, default) or implicit (e.g. "Firmicutes" for all ranks below Phylum).}
}
\value{
physeq object with correct name
}
\description{
Spread taxonomy according to last known taxa, to remove unknown and multi-affiliations
}
\examples{
library(phyloseq)
physeq <- phyloseq(otu_table(matrix(1:4, 2, 2), taxa_are_rows = TRUE),
tax_table(matrix(c("Firmicutes", "Firmicutes", "Unknown", "Bacilli", NA, "Lactobacillales"), 2, 3)))
tax_spread(physeq) |> tax_table()
tax_spread(physeq, explicit = FALSE) |> tax_table()
}
