% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{write_biom_safe}
\alias{write_biom_safe}
\title{Safe replacement of \code{\link[biomformat:write_biom]{biomformat::write_biom()}}}
\usage{
write_biom_safe(x, biom_file)
}
\description{
Works for biom data with only one column (sample), unlike the original version.
}
\examples{
library(phyloseq)
data(food)
small_food <- food |>
  subset_samples(EnvType == "DesLardons") |>
  subset_taxa(Genus == "Serratia")
small_biom <- small_food |> phyloseq_to_biom()
tmp_biom <- tempfile()
write_biom_safe(small_biom, tmp_biom)
readLines(tmp_biom, warn = FALSE)
}
